Promoted LatRange and LonRange to types, introduced DatetimeRange
git-svn-id: https://svn.smr.arpa.emr.it/dballe/dballe/trunk@4792 5c06e07b-eeef-0310-a3fe-9f68d50baefc

+++ dballe/core/test-utils-core.h
@@ -146,7 +146,7 @@ struct test_group : public tut::group_base
     tut::test_result run_test(int n) override
     {
         --n; // From 1-based to 0-based
-        if (n < 0 || n >= tests.size())
+        if (n < 0 || (size_t)n >= tests.size())
             throw tut::beyond_last_test();
 
         if (!test_can_run(name, tests[n].name))
